Fixes suffocation damage and healing for reagents (#74292)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This fixes a hidden side effect of the "refactor" in #73147 that
actually caused ALL existing suffocation damage dealing/healing
chemicals to not work on species that breathe gases other than oxygen.

This is also good for future code since if you want a chemical to not
affect a specific species or such, you should explicitly define that, it
shouldn't be implicitly done for you in the background with 0
implication.

Also reverts #74247 since it's no longer needed.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

First off, the change was unintended. It was intended to be a refactor.

Second, the change is unfun and isn't explained anywhere. It isn't very
fun to sit on the floor in medbay while even the most experienced
doctors can't figure out why the strongest suffocation healing chemicals
in the game aren't working. Cue this on repeat for 30 minutes. I've
personally experienced this like 10 times now and even I didn't know why
it was happening.

Even if you know the change exists, it's still extremely convoluted to
get around. You have to do CPR on a plasmaman at the speed of an average
cheetah or abuse the brute/burn revival method.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: Suffocation-handling reagents work on all species once more.
spellcheck: Added a previous tip for handling plasmaman suffocation back
into the game.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->