Fixes Ceiling Generation for Lavaland Icemoon Ruin - Whoops (#74567)
## About The Pull Request

Forgot to do this in #74410. It's **not** the reason why we see active
turfs on icebox station though, because this is just closed turfs. The
active turfs never spawned as a result of this, they only spawn as a
result of the open turfs of the ruin mixing with chasm air. The ceiling
looks a lot better now though.
## Why It's Good For The Game

Should have always been like this.
## Changelog
doesn't really matter.