Piercing rounds won't infinitely recurse when fired at piggyback players (#74586)
## About The Pull Request

I noticed in a round that someone fired a honkmother staff at a guy
riding a cyborg and it reported hitting them 374 times (with no effect).
Running it on local it seems like this proc was infinitely recursing
because after validating its targets it would return the mob the target
was buckled to, rather than the mob we actually validated, and then just
do that over and over again.
I made it not do that.

Now if you shoot someone riding a cyborg with a honk staff they will
fall of, as god intended.
Also now if you shoot a pair of people who are riding piggyback with
penetrator rounds, it will do what the name says and shoot both of them.

## Why It's Good For The Game

Rutimes bad. Throwing people around with bananas good. Shooting a hole
through two guys at a time, also good.

## Changelog

:cl:
fix: You can no longer escape the wrath of the honkmother by climbing
onto a cyborg.
fix: Shooting a pair of piggybacked players with a penetrator round will
penetrate both players.
/:cl: