Shelter in a Sands: Underground Lairs. (#2356)
* Shelter in a Sands: Underground Lairs.

Added a few underground lairs that could be found on a map.

Remade an "escape way of Quarry", made an Old Mine closed, with only one entrance on a surface.

Added a basement for Motel. As I wanted in a first time.

Added a "drafted version" of Deathclaw Lair. It's a radscorpions nest for now.

Shifted one of the buildings in Barony closer to other.  Remade un upper z-level area nearby of it a bit.

Replaced some walls in a buildings. Yeah, Original, it;s where siding walls MUST be use in a first.

* Fixes mining lamps, touches up loot

---------

Co-authored-by: Edward Nashton <eddienigma48@gmail.com>
Co-authored-by: Professor Popoff <omniderpectional@gmail.com>