Map based jobs and related tweaks (#2451)
Finally cracks the code of map based jobs, they specifically use the _end_ of the path that a job has. Which in the PR that the .json job config was added in was pointed out by Rohesie as potentially causing issues down the line but it seems like it made it through regardless so here we are! Some job paths might be changed due to this, like Ranger Doctor getting a slightly different end path to not conflict with Snowcrest Doctor.

Also did most of the changes from #2426 that were global, like disabling old NCR and Town roles.