Sulfur and Dust: New Faction Base and Ore Refinery (#2479)
* Sulfur and Dust: New Faction Base and Ore Refinery

Entirely replaced Ranger's Prison on a base of new faction (Dusters or whatever they will be named). That area was a some kind of living and administration area of Mining Complex previously, and has a Living Quarters and Medbay. Administration building is abandoned, but could be cleared and used for any Duster's purposes.
Sheriff's (or Leader's/Boss') Quarters + Prison + Armory was made by the faction using an adobe.

They also built a wall to separate themselves from the rest of the mine. There are small tunnel under the wall for those brave ones who decided to raid a Mine.

I also updated a Mine - added a new Entrance Building (like a checkpoint with a gear), and Ore Refinery. Also, the small area of Barony was changed (that one near "death pit").

Prison and old building from Mining Area were removed to Purgatory.

* roof helpers and minor tweaks

---------

Co-authored-by: Hekzder <masontrt@gmail.com>