Steel and Bloodbath: New Sentry Bot Area (#2498)
* Steel and Bloodbath: New Sentry Bot Area

Made a new area for sentry bot. Now it's entire robotic workshop with a two floors and basement. And yeah, you still need to deal with a bot to reach a good loot. Good luck with that.

* NEw Boss Area: F*ck this, we ball edition

Yeah, I've updated the entire damn area because I didn't like an old variant.

Sent a second floor of old area in a purgatory. Also upgraded a Junkyard PRess, upper level.

I hope it will be commited normally. Bless RNG.

* fixes missing quotation marks

* More more...

* Mobs, loot, FIXES. hooooly hecc

* slaughters random paste chunks/space

* aaahhh

* the last???

* THE LAST MISSING QUOTATION...

* another random pasted chunk; floating into the abyss Once strong now ded

---------

Co-authored-by: Professor Popoff <omniderpectional@gmail.com>