New ladders (#2505)
This pull request adds a bunch of ladders that had sprites but were not implemented:

Rope: just like a ladder, yet slower to climb up than to go down.
Bunker hatch: like a ladder, but the top can be closed or opened. You can even lock it from above using the hand-wheel.
Enclave bunker hatch: like a ladder, you can close or open it, but not lock it.