The Gravening (#2511)
This PR adds graves and tombstones (cairn, slab and wooden cross) to wake up your inner Undertaker.

Grave:

Just click on a desert turf with a shovel in hand to create the outlines of a grave.
You can then fill or dig the grave by using your shovel standing next to the grave (cannot work on it if you are on the turf or if you are incapacitated or resting).
You can bury 20 entities at max, with 1 human-sized mob.
LMB dig the grave, RMB fill the grave.
You can bury people alive and enjoy their muffled screams for help!
When someone is buried, they cannot breath anymore and take oxy damages. You can try to escape the grave if there is only 1 or 2 levels of dirt on top of you, if you make it in time before suffocating.
If you walk on an open grave, you will fall in it and badly hurt yourself.
You can push and shove people into graves.
Tombstones:

You can engrave messages on tombstones with a knife.
Slab can be made from 2 ceramic, cairn can be made from 1 ceramic
Wooden plank can be made with 1 plank, wooden cross from 2 planks
You can right-click on slabs and crosses to rotate/change the sprites.